+++ Assets/PurrNet/Runtime/Utils/ClonesContext.cs
@@ -1,4 +1,6 @@
+using System;
 using System.IO;
+using UnityEngine;
 
 namespace PurrNet.Utils
 {
@@ -26,5 +28,32 @@ namespace PurrNet.Utils
 #endif
             }
         }
+
+#if UNITY_EDITOR
+        public static string GetOriginalProjectPath()
+        {
+            if (!isClone)
+                return null;
+            try
+            {
+#if NET6_0_OR_GREATER
+                var target = new DirectoryInfo("Assets").ResolveLinkTarget(true);
+                if (target != null)
+                    return Path.GetDirectoryName(target.FullName);
+#else
+                var projectRoot = Path.GetDirectoryName(Application.dataPath) ?? ".";
+                var folderName = Path.GetFileName(projectRoot.TrimEnd(Path.DirectorySeparatorChar, '/', '\\'));
+                var cloneIdx = folderName.IndexOf("_clone", StringComparison.OrdinalIgnoreCase);
+                if (cloneIdx > 0)
+                {
+                    var parentName = folderName.Substring(0, cloneIdx);
+                    return Path.Combine(Path.GetDirectoryName(projectRoot) ?? "", parentName);
+                }
+#endif
+            }
+            catch { }
+            return null;
+        }
+#endif
     }
 }